split is_email tests to their own file
so devs can run portions of the tests
